telegram/cat_bread_bot/requirements.txt
torch==1.9.0
Pillow==9.4.0
telebot
numpy

